pyproject.toml


[tool.poetry]
name = "moments-cdk-app"
version = "0.1.0"
description = "My Moments CDK app"
authors = ["iamjustinthecloud <iamjustinthecloud@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
boto3 = "^1.40.45,<2.0.0"
attrs = ">=25.3.0,<26.0.0"
google-api-python-client = "^2.185.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
aws-lambda-powertools = "^3.23.0"
cdk-nag = "^2.37.55"


[tool.pyright]
typeCheckingMode = "basic"
reportMissingImports = true
reportUnknownArgumentType = "none"
reportCallIssue = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"
reportPrivateImportUsage = "none"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
pytest-snapshot = "^0.9.0"
colorlog = "^6.9.0"
pytest-sugar = "^1.1.1"
mypy-boto3-dynamodb = "^1.40.56"
boto3-stubs = {extras = ["essential"], version = "^1.40.59"}
aws-cdk-lib = ">=2.219.0,<3.0.0"
constructs = "^10.4.2,<11.0.0"
aws-cdk-aws-lambda-python-alpha = "^2.221.1a0"
mdutils = ">=1.8.0,<2.0.0"
isort = "^7.0.0"
coverage = "^7.11.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"

[tool.black]
line-length = 90
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 90
known_first_party = ["moments_app", "lambdas"]
combine_as_imports = true
skip_gitignore = true
ensure_newline_before_comments = true
